basic rendering loop setup, need to code server

+++ server/src/structs/player.rs
@@ -1,15 +1,15 @@
 //pub mod Player {
 
-
+#[derive(serde::Serialize)]
 pub struct Player {
     pub uuid: i32,
     pub name: String,
-    pub x: f64,
-    pub y: f64,
-    pub vx: f64,
-    pub vy: f64,
-    pub lx: f64,
-    pub ly: f64,
+    pub x: i32,
+    pub y: i32,
+    pub vx: i32,
+    pub vy: i32,
+    pub lx: i32,
+    pub ly: i32,
     pub scale: i32
 }
 
@@ -18,23 +18,23 @@ impl Player {
         Self {
             uuid,
             name: String::from("unknown"),
-            x: 0.0,
-            y: 0.0,
-            vx: 0.0,
-            vy: 0.0,
-            lx: 0.0,
-            ly: 0.0,
+            x: 0,
+            y: 0,
+            vx: 0,
+            vy: 0,
+            lx: 0,
+            ly: 0,
             scale
         }
     }
 
-    pub fn update_pos(&mut self, x: f64, y: f64) -> &mut Self {
+    pub fn update_pos(&mut self, x: i32, y: i32) -> &mut Self {
         self.x = x;
         self.y = y;
         self
     }
 
-    pub fn update_last_pos(&mut self, x: f64, y: f64) -> &mut Self {
+    pub fn update_last_pos(&mut self, x: i32, y: i32) -> &mut Self {
         self.lx = x;
         self.ly = y;
         self
